refactor: define an abstract form and cloud repository
This change should fix duplication code smell and make more testable.

TISNEW-5376: Show unsubmitted forms
